added generic upload ability

+++ npm/npm.go
@@ -35,7 +35,7 @@ type Metadata struct {
 }
 
 //GetNPMMetadata blah
-func GetNPMMetadata(creds auth.Creds, URL, packageIndex, packageName, configPath string, dlFolder string, workerNum int) {
+func GetNPMMetadata(creds auth.Creds, URL, packageIndex, packageName, configPath string, dlFolder string, workerNum int, flags helpers.Flags) {
 	data, _, _ := auth.GetRestAPI("GET", true, URL+packageName, creds.Username, creds.Apikey, "", nil, 1)
 	var metadata = artifactMetadata{}
 	err := json.Unmarshal([]byte(data), &metadata)
@@ -43,22 +43,24 @@ func GetNPMMetadata(creds auth.Creds, URL, packageIndex, packageName, configPath
 		log.Error("Worker ", workerNum, " error:"+err.Error())
 	}
 	for i, j := range metadata.Versions {
-		packageDl := packageIndex + "-" + i + ".tgz"
 
 		//TODO can be a problem if you set override base url and the tarball URL no longer matches correctly with creds.URL due to overwrite-base-url header
 		s := strings.Split(j.Dist.Tarball, URL)
 		//fmt.Println(len(s), "length of s") //413 error
 		if len(s) > 1 && s[1] != "" {
-			_, headStatusCode, _ := auth.GetRestAPI("HEAD", true, creds.URL+"/"+creds.Repository+"-cache/"+s[1], creds.Username, creds.Apikey, "", nil, 1)
+			_, headStatusCode, _ := auth.GetRestAPI("HEAD", true, creds.URL+"/"+flags.RepoVar+"-cache/"+s[1], creds.Username, creds.Apikey, "", nil, 1)
 			if headStatusCode == 200 {
-				log.Debug("Worker ", workerNum, " skipping, got 200 on HEAD request for ", creds.URL+"/"+creds.Repository+"-cache/"+s[1])
+				log.Debug("Worker ", workerNum, " skipping, got 200 on HEAD request for ", creds.URL+"/"+flags.RepoVar+"-cache/"+s[1])
 				continue
 			}
 		}
-		log.Info("Worker ", workerNum, " Downloading ", packageIndex, i, j.Dist.Tarball)
-		auth.GetRestAPI("GET", true, j.Dist.Tarball, creds.Username, creds.Apikey, configPath+dlFolder+"/"+packageDl, nil, 1)
-		err2 := os.Remove(configPath + dlFolder + "/" + packageDl)
-		helpers.Check(err2, false, "Deleting file", helpers.Trace())
+		if !flags.NpmMetadataVar {
+			packageDl := packageIndex + "-" + i + ".tgz"
+			log.Info("Worker ", workerNum, " Downloading ", s[1])
+			auth.GetRestAPI("GET", true, j.Dist.Tarball, creds.Username, creds.Apikey, configPath+dlFolder+"/"+packageDl, nil, 1)
+			err2 := os.Remove(configPath + dlFolder + "/" + packageDl)
+			helpers.Check(err2, false, "Deleting file", helpers.Trace())
+		}
 	}
 	helpers.Check(err, false, "Reading", helpers.Trace())
 	if err != nil {
